Fix parsing of 'show protocols' with bird 1.4
The current parsing was broken because of a change in the date format.

Actually, the new method is much simpler, and should resist small syntax
changes in Bird's output.  We don't use an ugly regexp anymore.

Important limitation: parsing will be messed up if the date contains a
space character.  This does not happen with the default date format of
both bird 1.3 and bird 1.4, but since the date format is configurable in
bird, it may happen anyway.